{"types":"C","runs":6}

+++ Problem_17.java
@@ -5,7 +5,7 @@ public class Problem_17 extends ConsoleProgram {
 	public void run(){
 		double x = readDouble("enter your nymber: ");
 		println((int) x);
-		println(x - (double) x);
+		println(x - (int) x);
 		
 	}
 }
